# docs/04_api/openapi.yaml
openapi: 3.0.3
info:
  title: Trading Lab API
  version: 1.0.0
  description: REST API contract for the Trading Lab frontend.
servers:
  - url: /api/v1
paths:
...
  /experiments:
    post:
      summary: Create experiment
      operationId: createExperiment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExperimentRequest'
      responses:
        '201':
          description: Experiment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateExperimentResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List experiments
      operationId: listExperiments
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/ExperimentStatus'
        - in: query
          name: strategyType
          schema:
            $ref: '#/components/schemas/StrategyType'
        - in: query
          name: mode
          schema:
            $ref: '#/components/schemas/ExperimentMode'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Experiment summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExperimentSummaryResponse'
...
  /experiments/{experimentId}/orders:
    get:
      summary: List experiment orders
      operationId: listExperimentOrders
      parameters:
        - $ref: '#/components/parameters/ExperimentId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Experiment orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrderResponse'
        '404':
          description: Experiment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /experiments/{experimentId}/trades:
    get:
      summary: List experiment trades
      operationId: listExperimentTrades
      parameters:
        - $ref: '#/components/parameters/ExperimentId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Experiment trades
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTradeResponse'
        '404':
          description: Experiment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /experiments/{experimentId}/broker-sync-logs:
    get:
      summary: List experiment broker sync logs
      operationId: listExperimentBrokerSyncLogs
      parameters:
        - $ref: '#/components/parameters/ExperimentId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Experiment broker sync logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBrokerSyncLogResponse'
        '404':
          description: Experiment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /experiments/{experimentId}/paper-status:
    get:
      summary: Get paper trading operational status
      operationId: getExperimentPaperStatus
      parameters:
        - $ref: '#/components/parameters/ExperimentId'
      responses:
        '200':
          description: Paper trading operational status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaperStatusResponse'
        '404':
          description: Experiment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    ExperimentId:
      name: experimentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 500
...
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
  schemas:
...
    ExperimentMode:
      type: string
      enum: [HISTORICAL_SIMULATION, LIVE_SIMULATION, PAPER_TRADING]
    StrategyType:
      type: string
      enum: [BUY_AND_HOLD, MOVING_AVERAGE, AGENTIC_AI, OPENING_RANGE_BREAKOUT]
    ExperimentStatus:
      type: string
      enum: [CREATED, RUNNING, PAUSED, STOPPED, COMPLETED, FAILED]
    TradingFrequency:
      type: string
      enum: [DAILY, WEEKLY, MONTHLY, INTRADAY_5_MIN]
    FeeModelType:
      type: string
      enum: [NONE, FIXED, PERCENTAGE]
    ExecutionStepStatus:
      type: string
      enum: [RUNNING, COMPLETED, SKIPPED, FAILED]
    TriggerType:
      type: string
      enum: [HISTORICAL, SCHEDULED, MANUAL]
...
    TradeAction:
      type: string
      enum: [BUY, SELL, HOLD]
    FinalAction:
      type: string
      enum: [BUY, SELL, HOLD]
    OrderMode:
      type: string
      enum: [SIMULATED, PAPER_BROKER]
    BrokerName:
      type: string
      enum: [ALPACA, NONE]
    OrderSide:
      type: string
      enum: [BUY, SELL]
    OrderType:
      type: string
      enum: [MARKET]
    OrderStatus:
      type: string
      enum: [CREATED, SUBMITTED, FILLED, REJECTED, FAILED, CANCELLED]
    AgentMode:
      type: string
      enum: [SINGLE_AGENT, PIPELINE]
...
    BrokerSyncStatus:
      type: string
      enum: [SUCCESS, FAILED, MISMATCH]
...
    CreateExperimentRequest:
      type: object
      required: [name, mode, strategyType, assetSymbol, initialCapital, startDate, endDate, tradingFrequency, feeModelType, feeValue, strategyConfig]
      properties:
        name:
          type: string
        mode:
          $ref: '#/components/schemas/ExperimentMode'
        strategyType:
          $ref: '#/components/schemas/StrategyType'
        assetSymbol:
          type: string
          example: SPY
        initialCapital:
          type: number
          format: double
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        tradingFrequency:
          $ref: '#/components/schemas/TradingFrequency'
        feeModelType:
          $ref: '#/components/schemas/FeeModelType'
        feeValue:
          type: number
          format: double
        strategyConfig:
          $ref: '#/components/schemas/StrategyConfigInput'
    StrategyConfigInput:
      type: object
      required: [strategyVersion]
      properties:
        strategyVersion:
          type: string
        movingAverageWindow:
          type: integer
          nullable: true
        positionSizingType:
          type: string
          enum: [ALL_IN, FIXED_CASH, PERCENT_OF_PORTFOLIO, FIXED_QUANTITY]
          nullable: true
        positionSizingValue:
          type: number
          format: double
          nullable: true
        agentMode:
          allOf:
            - $ref: '#/components/schemas/AgentMode'
          nullable: true
        modelName:
          type: string
          nullable: true
        confidenceThreshold:
          type: number
          format: double
          nullable: true
        parametersJson:
          type: object
          additionalProperties: true
          nullable: true
          example:
            riskConfig:
              maxPositionSizePct: 1.0
              maxTradesPerDay: null
              maxTradesPerWeek: null
              maxDrawdownPct: null
              drawdownAction: BLOCK_TRADES
              fallbackAction: HOLD
    Experiment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        mode:
          $ref: '#/components/schemas/ExperimentMode'
        strategyType:
          $ref: '#/components/schemas/StrategyType'
        assetSymbol:
          type: string
        status:
          $ref: '#/components/schemas/ExperimentStatus'
        initialCapital:
          type: number
          format: double
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        tradingFrequency:
          $ref: '#/components/schemas/TradingFrequency'
        feeModelType:
          $ref: '#/components/schemas/FeeModelType'
        feeValue:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Portfolio:
      type: object
      properties:
        id:
          type: integer
          format: int64
        experimentId:
          type: integer
          format: int64
        cash:
          type: number
          format: double
        positionSymbol:
          type: string
          nullable: true
        positionQuantity:
          type: number
          format: double
          nullable: true
        currentPrice:
          type: number
          format: double
          nullable: true
        currentPositionValue:
          type: number
          format: double
          nullable: true
        currentPortfolioValue:
          type: number
          format: double
          nullable: true
        updatedAt:
          type: string
          format: date-time
    CreateExperimentResponse:
      type: object
      properties:
        experiment:
          $ref: '#/components/schemas/Experiment'
        portfolio:
          $ref: '#/components/schemas/Portfolio'
    ExperimentSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        mode:
          $ref: '#/components/schemas/ExperimentMode'
        strategyType:
          $ref: '#/components/schemas/StrategyType'
        assetSymbol:
          type: string
        status:
          $ref: '#/components/schemas/ExperimentStatus'
        currentPortfolioValue:
          type: number
          format: double
          nullable: true
        totalReturn:
          type: number
          format: double
          nullable: true
        profitLoss:
          type: number
          format: double
          nullable: true
        numberOfTrades:
          type: integer
          nullable: true
        maxDrawdown:
          type: number
          format: double
          nullable: true
        lastTrade:
          allOf:
            - $ref: '#/components/schemas/TradeSummary'
          nullable: true
        latestAgentDecisions:
          type: array
          items:
            $ref: '#/components/schemas/AgentDecisionSummary'
    PaginatedExperimentSummaryResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponseBase'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ExperimentSummary'
...
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        executionStepId:
          type: integer
          format: int64
        experimentId:
          type: integer
          format: int64
        riskCheckId:
          type: integer
          format: int64
        mode:
          $ref: '#/components/schemas/OrderMode'
        brokerName:
          $ref: '#/components/schemas/BrokerName'
        brokerOrderId:
          type: string
          nullable: true
        symbol:
          type: string
        side:
          $ref: '#/components/schemas/OrderSide'
        quantity:
          type: number
          format: double
        orderType:
          $ref: '#/components/schemas/OrderType'
        status:
          $ref: '#/components/schemas/OrderStatus'
        submittedAt:
          type: string
          format: date-time
          nullable: true
        filledAt:
          type: string
          format: date-time
          nullable: true
        averageFillPrice:
          type: number
          format: double
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    Trade:
      type: object
      properties:
        id:
          type: integer
          format: int64
        executionStepId:
          type: integer
          format: int64
        experimentId:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time
        symbol:
          type: string
        side:
          $ref: '#/components/schemas/OrderSide'
        quantity:
          type: number
          format: double
        price:
          type: number
          format: double
        orderValue:
          type: number
          format: double
          nullable: true
        fee:
          type: number
          format: double
          nullable: true
        portfolioValueAfterTrade:
          type: number
          format: double
          nullable: true
        createdAt:
          type: string
          format: date-time
    TradeSummary:
      type: object
      properties:
        side:
          $ref: '#/components/schemas/OrderSide'
        quantity:
          type: number
          format: double
        price:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
...
    AgentDecisionSummary:
      type: object
      properties:
        executionStepId:
          type: integer
          format: int64
        agentMode:
          $ref: '#/components/schemas/AgentMode'
        action:
          $ref: '#/components/schemas/TradeAction'
        confidence:
          type: number
          format: double
          nullable: true
        reason:
          type: string
          nullable: true
        finalAction:
          $ref: '#/components/schemas/FinalAction'
        createdAt:
          type: string
          format: date-time
...
    BrokerSyncLog:
      type: object
      properties:
        id:
          type: integer
          format: int64
        executionStepId:
          type: integer
          format: int64
        experimentId:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time
        brokerName:
          $ref: '#/components/schemas/BrokerName'
        syncStatus:
          $ref: '#/components/schemas/BrokerSyncStatus'
        brokerCash:
          type: number
          format: double
          nullable: true
        localCash:
          type: number
          format: double
          nullable: true
        brokerPositionsJson:
          type: object
          additionalProperties: true
          nullable: true
        localPositionsJson:
          type: object
          additionalProperties: true
          nullable: true
        mismatchDetailsJson:
          type: object
          additionalProperties: true
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    PaperExecutionStepSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/ExecutionStepStatus'
        triggerType:
          $ref: '#/components/schemas/TriggerType'
        sequenceNumber:
          type: integer
        scheduledFor:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    PaperStatusResponse:
      type: object
      properties:
        experimentId:
          type: integer
          format: int64
        experimentStatus:
          $ref: '#/components/schemas/ExperimentStatus'
        mode:
          $ref: '#/components/schemas/ExperimentMode'
        strategyType:
          $ref: '#/components/schemas/StrategyType'
        tradingFrequency:
          $ref: '#/components/schemas/TradingFrequency'
        assetSymbol:
          type: string
        supportedByPaperScheduler:
          type: boolean
        paperTradingSchedulerEnabled:
          type: boolean
        alpacaPaperTradingEnabled:
          type: boolean
        dailyEvaluationTime:
          type: string
        timezone:
          type: string
        currentDueSlot:
          type: string
          format: date-time
          nullable: true
        nextEligibleEvaluationTime:
          type: string
          format: date-time
          nullable: true
        alreadyExecutedCurrentDueSlot:
          type: boolean
        openSubmittedOrdersCount:
          type: integer
        lastBrokerSyncTimestamp:
          type: string
          format: date-time
          nullable: true
        lastPaperExecutionStep:
          allOf:
            - $ref: '#/components/schemas/PaperExecutionStepSummary'
          nullable: true
        reasonCode:
          type: string
        message:
          type: string
...
    PaginatedResponseBase:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
    PaginatedTradeResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponseBase'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Trade'
    PaginatedOrderResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponseBase'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Order'
...
    PaginatedBrokerSyncLogResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponseBase'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/BrokerSyncLog'
    ErrorResponse:
      type: object
      required: [errorCode, message, details]
      properties:
        errorCode:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true